rename pv costing model file for consistency
